backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
motor==3.3.2
python-dotenv==1.0.0
openai==1.12.0
httpx==0.26.0
pydantic==2.6.1